# WARNING: head commit changed in the meantime
Merge branch 'master' of
https://github.com/jaiganeshsha/SDET34Framework.git

+++ java/com/crm/comcast/genericUtility/Excelutility.java
@@ -51,7 +51,7 @@ public class Excelutility {
     	Workbook wb = WorkbookFactory.create(fis);
     	Sheet sh = wb.getSheet(sheet);
     	int lastrow = sh.getLastRowNum();
-    	short lastcell = sh.getRow(0).getLastCellNum();
+    	int lastcell = sh.getRow(0).getLastCellNum();
     	Object[][] data=new Object[lastrow][lastcell];
     	for(int i=0;i<lastrow;i++) {
     		for(int j=0;j<lastcell;j++) {
